feat!: vin api, extracted functions  - replaced vin api  - updated functionality to acommodate new vin api  - updated env.example  - extracted functionality, api functionality to individual files
